requirements.txt
# Install PyTorch supporting cuda version 11080
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.1+cu118
torchvision==0.17.1+cu118
torchaudio==2.2.1+cu118
soundfile==0.12.1

numpy==1.23.5
matplotlib==3.8.3

tensorflow==2.12.0
ray[tune]==2.38.0
optuna==4.1

partitura==1.6.0
PyYAML==6.0.2